Fix init sequence so correct GPIO pins are initialized AND used
Without this patch there is the possibility that upon initilization solard uses default GPIO pins, and after setting pins direction, reads config only to have set the direction of wrong pins. This commit fixes the order of init sequence so that pin numbers are first read, and then initialized for use.