FIX: Startup behavoir - failure to recover from wrong sensor data
Expand the time where ANY values for sensors are accepted.
The reasoning for this is that on a system that is starting up, the load is initially high, and I've had a case when furnace temp was read wrong, and solard failed to recover.
This commit commit hopefully fixes such cases by just lifting its sensor-data protections for longer, thus hoping that for last any-sensor-data the system will have settled, and the values are real.